+++ desktop/js/veralink.js
@@ -27,8 +27,10 @@ $("#table_cmd").sortable({
 });
 
 /* Fonction permettant l'affichage de l'équipement */
-function printEqLogic(data) {
-  //console.log(data);
+function printEqLogic(eqLogic) {
+  // TODO get the proper icon from eqlogic and set it into the IMG attribute
+  //console.log(eqLogic);
+  //$('#icon_visu').attr('src', 'plugins/reolink/desktop/img/camera' + _eqLogic.id + '.png')
 }
 
 /* Fonction permettant l'affichage des commandes dans l'équipement */