+++ docs/fr_FR/changelog.md
@@ -4,6 +4,10 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 07/05/2023
+
+- when saving the vera root Equipment, it will remove from Jeedom the equipments for VERA devices that do not exist anymore
+
 # 30/08/2022
 
 - Add Jeedom Info Command "Alarme mode" to security sensors ( motion, door, humidity ) to report the 'Armed' variable state from vera devices